Works, added example, fixed docs
